Added code and functionality to upload cleaned and generated features to S3 bucket
