+++ data/auxiliary/new_ipo_data.csv
@@ -1,0 +1,21 @@
+,year,no_ipos
+0,2019,41
+1,2018,228
+2,2017,186
+3,2016,107
+4,2015,173
+5,2014,261
+6,2013,230
+7,2012,146
+8,2011,143
+9,2010,165
+10,2009,63
+11,2008,50
+12,2007,279
+13,2006,240
+14,2005,236
+15,2004,249
+16,2003,84
+17,2002,85
+18,2001,96
+19,2000,256
