+++ src/clean_acquisitions_data.py
@@ -5,8 +5,9 @@ import argparse
 
 import yaml
 import pandas as pd
+import boto3
 
-import src.helpers.gen_helpers as gen_h
+from helpers import gen_helpers as gen_h
 
 logging.basicConfig(format="%(asctime)s - %(name)s - %(levelname)s - %(message)s")
 logger = logging.getLogger(__name__)
@@ -15,6 +16,7 @@ def get_acquisition_count(df, all_companies_file):
     """Aggregates acquisition by acquirer and creates new dataframe for
     acquisition count"""
     # Create list of acquired companies in original company dataset
+    company_df = gen_h.read_data(all_companies_file)
     all_companies_list = list(company_df['permalink'].unique())
     acquired_companies_list = list(df['company_permalink'].unique())
     acquirer_companies_list = list(df['acquirer_permalink'].unique())
@@ -49,10 +51,14 @@ def run_clean_acquisitions(args):
         config = yaml.load(f)
 
     df = gen_h.read_data(args.input_file)
-    df = gen_acquisition_count(df, config['clean_acquisitions_data']
+    df = get_acquisition_count(df, **config['clean_acquisitions_data']
     ['get_acquisition_count'])
 
     df.to_csv(args.save)
+
+    # Save copy to S3 Bucket
+    s3 = boto3.client("s3")
+    s3.upload_file(args.save, args.bucket_name, args.output_file_path)
 
 if __name__ == '__main__':
     parser = argparse.ArgumentParser(description="")
@@ -60,6 +66,10 @@ if __name__ == '__main__':
     parser.add_argument('--input_file', help='path to csv file with raw data')
     parser.add_argument('--save', default='data/auxiliary/new_acquisitions_data.csv',
     help='path to where the cleaned dataset should be saved to')
+    parser.add_argument("--bucket_name", default='startup-funding-working-bucket',
+    help="S3 bucket name")
+    parser.add_argument("--output_file_path", default='new_acquisitions_data.csv',
+    help="output file path of uploaded file")
 
     args = parser.parse_args()
